requirements.txt
streamlit
openai
gitpython